feat: True TBR two-pass hardware binning + BorgSize refactor + poster
Step 31: Multi-Triangle Autonomous Rendering
- 128-byte descriptor layout with tile-aligned bbox
- Hardware-driven IMEM staging for rast/frag shaders
- Sequential multi-triangle loop with triCount=0 guard
- Hardware bbox walk with full tile iteration FSM

Step 32: True TBR Two-Pass Hardware Binning
- BorgBinner: new Chisel module writing triangle indices to per-tile
  bin lists in PSRAM; 5/5 unit tests green
- BorgSequencer: geometry pass (Pass 1) bins all triangles first;
  tile render pass (Pass 2) flushes each tile exactly once
- 33 FSM states; two new MMIO regs (SEQ_BIN_BASE, SEQ_BIN_ROW_BYTES,
  SEQ_SETUP_BASE); DMA arb unified as geo engine
- Driver: borgBinRenderAutonomous() writes 7 MMIO regs + trigger,
  ~16K CPU PSRAM writes eliminated
- TBR_SETUP_ENTRY_BYTES: 64→128; docs/07_tbr.md added

Refactoring:
- Replace 5 BorgConfig feature flags with BorgSize sealed trait
  (Small/Large), eliminating all nested match blocks in Borg.scala
- BorgConfig.PicoIce → BorgSize.Small; BorgConfig.Sim → BorgSize.Large

Documentation:
- Add HPG 2026 poster (docs/poster/poster.tex + abstract.tex)
- Update poster: Phase 2 complete, BorgBinner/Sequencer in milestones
- Add docs/poster/.gitignore; fix Makefile to use ../hw_diagram.svg
- Add docs/bibliography.bib

Environment:
- flake.nix: switch to texlive.combined.scheme-full for poster build

make test-all green (all suites pass)

+++ software/borg/borg_sys.h
@@ -40,8 +40,8 @@
 #ifndef PSRAM_BASE
 #define PSRAM_BASE 0x01001000
 #endif
-#define PSRAM_IN(n)   (*(volatile uint32_t *)(PSRAM_BASE + (n) * 4))
-#define PSRAM_OUT(n)  (*(volatile uint32_t *)(PSRAM_BASE + PSRAM_OUT_OFFSET + (n) * 4))
+#define PSRAM_IN(n)   (*(volatile uint32_t *)(uintptr_t)(PSRAM_BASE + (n) * 4))
+#define PSRAM_OUT(n)  (*(volatile uint32_t *)(uintptr_t)(PSRAM_BASE + PSRAM_OUT_OFFSET + (n) * 4))
 // Companion to PSRAM_OUT: converts the same word-index n to the raw SPI byte
 // address used by the hardware flusher (gpuMem.addr in MemoryController).
 // PSRAM_OUT(n) and PSRAM_OUT_SPI(n) address the same physical PSRAM word;
@@ -50,7 +50,7 @@
 // Raw SPI byte address accessor: reads/writes a 32-bit word at an absolute SPI
 // byte address.  Used by the sequencer path to access shader/descriptor regions
 // at fixed addresses (Step 29.5).
-#define PSRAM_OUT_RAW(spi_addr) (*(volatile uint32_t *)(PSRAM_BASE - PSRAM_SPI_BASE + (spi_addr)))
+#define PSRAM_OUT_RAW(spi_addr) (*(volatile uint32_t *)(uintptr_t)(PSRAM_BASE - PSRAM_SPI_BASE + (spi_addr)))
 
 // --- Peripheral base addresses ---
 // These come from the SoC address map (soc.rdl).
@@ -66,7 +66,7 @@
 #endif
 
 // --- UART accessor macros ---
-#define UART_TX      (*(volatile uint32_t *)(UART_BASE + 0x0))
-#define UART_STATUS  (*(volatile uint32_t *)(UART_BASE + 0x4))
-#define UART_BAUD    (*(volatile uint32_t *)(UART_BASE + 0x8))
+#define UART_TX      (*(volatile uint32_t *)(uintptr_t)(UART_BASE + 0x0))
+#define UART_STATUS  (*(volatile uint32_t *)(uintptr_t)(UART_BASE + 0x4))
+#define UART_BAUD    (*(volatile uint32_t *)(uintptr_t)(UART_BASE + 0x8))
